feat(lastfm): service worker fallback
